Bitbucket server - Move getRepoGitUrl into utils
