+++ lib/modules/platform/bitbucket-server/index.ts
@@ -1,4 +1,3 @@
-import url from 'url';
 import is from '@sindresorhus/is';
 import delay from 'delay';
 import JSON5 from 'json5';
@@ -10,13 +9,7 @@ import {
   REPOSITORY_NOT_FOUND,
 } from '../../../constants/error-messages';
 import { logger } from '../../../logger';
-import {
-  BranchStatus,
-  HostRule,
-  PrState,
-  VulnerabilityAlert,
-} from '../../../types';
-import type { GitProtocol } from '../../../types/git';
+import { BranchStatus, PrState, VulnerabilityAlert } from '../../../types';
 import type { FileData } from '../../../types/platform/bitbucket-server';
 import * as git from '../../../util/git';
 import { deleteBranch } from '../../../util/git';
@@ -158,41 +151,6 @@ export async function getJsonFile(
   return JSON.parse(raw);
 }
 
-function getRepoGitUrl(
-  repository: string,
-  info: BbsRestRepo,
-  opts: HostRule
-): string {
-  let cloneUrl = info.links.clone?.find(({ name }) => name === 'http');
-  if (!cloneUrl) {
-    // Http access might be disabled, try to find ssh url in this case
-    cloneUrl = info.links.clone?.find(({ name }) => name === 'ssh');
-  }
-
-  let gitUrl: string;
-  if (!cloneUrl) {
-    // Fallback to generating the url if the API didn't give us an URL
-    const { host, pathname } = url.parse(defaults.endpoint);
-    gitUrl = git.getUrl({
-      protocol: defaults.endpoint.split(':')[0] as GitProtocol,
-      auth: `${opts.username}:${opts.password}`,
-      host: `${host}${pathname}${
-        pathname.endsWith('/') ? '' : /* istanbul ignore next */ '/'
-      }scm`,
-      repository,
-    });
-  } else if (cloneUrl.name === 'http') {
-    // Inject auth into the API provided URL
-    const repoUrl = url.parse(cloneUrl.href);
-    repoUrl.auth = `${opts.username}:${opts.password}`;
-    gitUrl = url.format(repoUrl);
-  } else {
-    // SSH urls can be used directly
-    gitUrl = cloneUrl.href;
-  }
-  return gitUrl;
-}
-
 // Initialize BitBucket Server by getting base branch
 export async function initRepo({
   repository,
@@ -233,7 +191,12 @@ export async function initRepo({
       throw new Error(REPOSITORY_EMPTY);
     }
 
-    const gitUrl = getRepoGitUrl(config.repositorySlug, info, opts);
+    const gitUrl = utils.getRepoGitUrl(
+      config.repositorySlug,
+      defaults.endpoint,
+      info,
+      opts
+    );
 
     await git.initRepo({
       ...config,
